Begin refactor for extra stats as playerstats is too large of a file.
